+++ projects/ngx-location-picker/src/lib/components/ngx-location-picker.component.ts
@@ -210,6 +210,8 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
   private cursorOnLeaflet = false;
   /* Circle drawn when trackingposition to show accuracy of coordinate */
   private proximityCircle;
+  /* Circle drawn when trackingposition to show coordinate */
+  private proximityCenter;
   /* Timeout to reset clearwatch*/
   private clearWatchTimeoutId;
 
@@ -859,7 +861,9 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
   private showProximityCircle(coords: number[], accuracy: number): void {
     this.removeProximityCircle();
     this.proximityCircle = this.locationViewerService.L.circle(coords, accuracy, { opacity: 0 });
+    this. proximityCenter = this.locationViewerService.L.circle(coords, 15, { fillOpacity: 1, color: '#FFFFFF', fillColor: '#0057b7', className: 'proximity-center' });
     this.proximityCircle.addTo(this.leafletMap.map);
+    this.proximityCenter.addTo(this.leafletMap.map);
     this.leafletMap.setView(coords, 15);
   }
 
@@ -870,6 +874,11 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
     if (this.proximityCircle) {
       this.proximityCircle.remove();
       this.proximityCircle = null;
+    }
+
+    if (this.proximityCenter) {
+      this.proximityCenter.remove();
+      this.proximityCenter = null;
     }
   }