+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- Added additional circle to improve visibilty of proximity circle
 
 ## [5.4.0] - 2022-09-05